+++ MacPacker/CHANGELOG.md
@@ -1,6 +1,7 @@
 v0.6
 - feat: hit space to open internal preview
-- chore: add support for macOS 13 and 14
+- feat: setting to change the breadcrumb position
+- chore: add support for macOS 14
 
 v0.5
 - feat: multiple windows support